Refactor FTPDownloadManager
